Added another test moved fixtures to conftest.py
